Update fields in the V5 FADT.
Integrate some late changes in the V5 (ACPI 5.0) FADT within the
ACPI 5.0 specification. Includes
changes to 2 flags and adds the sleep control/status registers.
Increases the size of the FADT to 0x10C.

+++ source/include/acdisasm.h
@@ -135,7 +135,7 @@
 typedef const struct acpi_dmtable_info
 {
     UINT8                       Opcode;
-    UINT8                       Offset;
+    UINT16                      Offset;
     char                        *Name;
     UINT8                       Flags;
